Rename developer tab to api sandbox (#36)
* Rename DeveloperTab to ApiSandbox and refactor component structure


* Refactor: Move ServerStats and HalLink types to models, update imports


---------
