+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as CodeBlock } from './CodeBlock';
-export { default as DeveloperTab } from './DeveloperTab/DeveloperTab';
+export { default as ApiSandbox } from './ApiSandbox/ApiSandbox';
 export { default as DropdownSelector } from './DropdownSelector';
 export { default as ErrorBoundary } from './ErrorBoundary';
 export { default as InputField } from './InputField';
