+++ src/pages/ApiSandbox.tsx
@@ -1,6 +1,6 @@
 import { QueryOptions } from '@tanstack/react-query';
 import { getAllUsers } from '../api';
-import DeveloperTabComponent from '../components/DeveloperTab/DeveloperTab';
+import ApiSandboxComponent from '../components/ApiSandbox/ApiSandbox';
 import { useAxios } from '../hooks';
 
 const ApiSandbox = () => {
@@ -10,7 +10,7 @@ const ApiSandbox = () => {
     return (
         <div className='max-w-7xl mx-auto p-4'>
             <h1 className='text-2xl font-bold mb-4'>Developer Tab Page</h1>
-            <DeveloperTabComponent queryOptions={queryOptions} fullUrl={fullUrl} />
+            <ApiSandboxComponent queryOptions={queryOptions} fullUrl={fullUrl} />
         </div>
     );
 };
